+++ src/lib/utils/fs.ts
@@ -134,9 +134,13 @@ export function copySync(src: string, dest: string): void {
 }
 
 /**
- * Simpler version of `glob.sync` that only covers our use cases, only ever matching files, and ignoring node_modules.
+ * Simpler version of `glob.sync` that only covers our use cases, always ignoring node_modules.
  */
-export function glob(pattern: string, root: string): string[] {
+export function glob(
+    pattern: string,
+    root: string,
+    options?: { includeDirectories?: boolean }
+): string[] {
     const result: string[] = [];
     const mini = new Minimatch(normalizePath(pattern));
     const dirs: string[][] = [normalizePath(root).split("/")];
@@ -147,6 +151,16 @@ export function glob(pattern: string, root: string): string[] {
         for (const child of fs.readdirSync(dir.join("/"), {
             withFileTypes: true,
         })) {
+            if (
+                child.isFile() ||
+                (options?.includeDirectories && child.isDirectory())
+            ) {
+                const childPath = [...dir, child.name].join("/");
+                if (mini.match(childPath)) {
+                    result.push(childPath);
+                }
+            }
+
             if (child.isDirectory() && child.name !== "node_modules") {
                 const childPath = dir.concat(child.name);
                 if (
@@ -155,11 +169,6 @@ export function glob(pattern: string, root: string): string[] {
                     )
                 ) {
                     dirs.push(childPath);
-                }
-            } else if (child.isFile()) {
-                const childPath = [...dir, child.name].join("/");
-                if (mini.match(childPath)) {
-                    result.push(childPath);
                 }
             }
         }