+++ src/test/slow/entry-point.test.ts
@@ -1,18 +1,48 @@
+import { tempdirProject } from "@typestrong/fs-fixture-builder";
 import { deepStrictEqual as equal, ok } from "assert";
 import { join } from "path";
 import { Application, EntryPointStrategy, TSConfigReader } from "../..";
 
-const root = join(__dirname, "entry-points");
+const fixture = tempdirProject();
+fixture.addJsonFile("tsconfig.json", {
+    include: ["."],
+});
+fixture.addFile("index.ts", "export function fromIndex() {}");
+fixture.addFile("extra.ts", "export function extra() {}");
 
 describe("Entry Points", () => {
+    beforeEach(() => {
+        fixture.write();
+    });
+
+    afterEach(() => {
+        fixture.rm();
+    });
+
     const app = new Application();
-    const tsconfig = join(root, "tsconfig.json");
+    const tsconfig = join(fixture.cwd, "tsconfig.json");
     app.options.addReader(new TSConfigReader());
 
     it("Supports expanding existing paths", () => {
         app.bootstrap({
             tsconfig,
-            entryPoints: [root],
+            entryPoints: [fixture.cwd],
+            entryPointStrategy: EntryPointStrategy.Expand,
+        });
+
+        const entryPoints = app.getEntryPoints();
+        ok(entryPoints);
+        equal(
+            entryPoints.length,
+            2,
+            "There are two files, so both should be expanded"
+        );
+    });
+
+    it("Supports expanding globs in paths", () => {
+        app.bootstrap({
+            tsconfig,
+            entryPoints: [`${fixture.cwd}/*.ts`],
             entryPointStrategy: EntryPointStrategy.Expand,
         });
 
@@ -28,7 +58,7 @@ describe("Entry Points", () => {
     it("Supports resolving directories", () => {
         app.bootstrap({
             tsconfig,
-            entryPoints: [root],
+            entryPoints: [fixture.cwd],
             entryPointStrategy: EntryPointStrategy.Resolve,
         });